backend/requirements.txt
fastapi
uvicorn
python-dotenv
anthropic
faiss-cpu
sentence-transformers
pypdf
beautifulsoup4
httpx
python-docx
python-multipart
